change temperature to float

+++ src/main/kotlin/tech/sprytin/starter/chatgpt/dto/ChatCompletionRequest.kt
@@ -3,7 +3,7 @@ package tech.sprytin.starter.chatgpt.dto
 data class ChatCompletionRequest(
     val messages: List<ChatMessage>,
     val model: String = GPT3_5_TURBO,
-    val temperature: Int = 1,
+    val temperature: Float = 1f,
 ) {
     companion object {
         const val GPT3_5_TURBO = "gpt-3.5-turbo"
